# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ana Portfolio API
  description: >-
    API do portfólio: conteúdo público, editor autenticado, aparência e
    assistente.


    ### Como usar (passo a passo)

    1. Teste a API em **GET /api/v1/health**.

    2. Leia dados com os **GET** 🔓 (públicos) e anote o campo `revision`.

    3. Para alterar dados, faça login em **POST /api/v1/auth/login** e copie o
    `accessToken`.

    4. Clique em **Authorize** (topo da página), cole o token e confirme.

    5. Use os endpoints 🔒 (POST, PUT, DELETE) enviando `expectedRevision`: a
    **versão atual dos dados**, igual ao `revision` do último GET.


    ### Erros comuns

    - **400**: campo faltando ou em formato errado.

    - **401**: sem login ou token vencido.

    - **404**: o `id` não existe.

    - **409**: `expectedRevision` desatualizado. Faça um novo GET e tente de
    novo.
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    def-1:
      type: object
      description: Identidade e contatos exibidos no portfólio.
      required:
        - name
        - area
        - role
        - level
        - company
        - cep
        - address
        - location
        - email
        - whatsapp
        - github
        - linkedin
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
          description: Nome completo.
          example: Maria Exemplo
        area:
          type: string
          minLength: 1
          maxLength: 100
          description: Área de atuação.
          example: Qualidade de Software
        role:
          type: string
          minLength: 1
          maxLength: 80
          description: Cargo ou função.
          example: QA Analyst
        level:
          type: string
          minLength: 1
          maxLength: 40
          description: Nível de senioridade.
          example: Pleno
        company:
          type: string
          maxLength: 120
          description: Empresa atual (pode ficar vazio).
          example: Empresa Exemplo
        cep:
          type: string
          maxLength: 20
          description: CEP (pode ficar vazio).
          example: 00000-000
        address:
          type: string
          maxLength: 240
          description: Endereço (pode ficar vazio).
          example: Rua Exemplo, 123
        location:
          type: string
          minLength: 1
          maxLength: 160
          description: Cidade e estado.
          example: São Paulo, SP
        email:
          type: string
          format: email
          description: E-mail de contato.
          example: maria@exemplo.com
        whatsapp:
          type: string
          maxLength: 40
          description: WhatsApp com DDD (pode ficar vazio).
          example: +55 11 90000-0000
        github:
          type: string
          format: uri
          description: Link completo do GitHub.
          example: https://github.com/exemplo
        linkedin:
          type: string
          format: uri
          description: Link completo do LinkedIn.
          example: https://www.linkedin.com/in/exemplo
        photoUrl:
          type: string
          maxLength: 3000000
          description: "Foto (opcional): URL pública ou data URL de imagem. Para trocar só
            a foto, use PUT /api/v1/profile/photo."
          example: https://exemplo.com/foto.jpg
      title: Profile
      example:
        name: Maria Exemplo
        area: Qualidade de Software
        role: QA Analyst
        level: Pleno
        company: Empresa Exemplo
        cep: 00000-000
        address: Rua Exemplo, 123
        location: São Paulo, SP
        email: maria@exemplo.com
        whatsapp: +55 11 90000-0000
        github: https://github.com/exemplo
        linkedin: https://www.linkedin.com/in/exemplo
    def-2:
      type: object
      description: Texto principal da Home.
      required:
        - title
        - intro
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 300
          description: Título da Home.
          example: Qualidade de software com método e curiosidade
        intro:
          type: string
          minLength: 1
          maxLength: 3000
          description: Texto de apresentação.
          example: Sou QA e ajudo times a entregar produtos confiáveis.
      title: HeroContent
      example:
        title: Qualidade de software com método e curiosidade
        intro: Sou QA e ajudo times a entregar produtos confiáveis.
    def-3:
      type: object
      description: Texto da seção Sobre.
      required:
        - title
        - body
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 300
          description: Título da seção.
          example: Sobre mim
        body:
          type: string
          minLength: 1
          maxLength: 5000
          description: Texto completo da seção.
          example: Trabalho com testes funcionais, automação e análise de falhas.
      title: AboutContent
      example:
        title: Sobre mim
        body: Trabalho com testes funcionais, automação e análise de falhas.
    def-4:
      type: object
      description: Uma competência (habilidade) do portfólio.
      required:
        - id
        - title
        - description
        - tags
      additionalProperties: false
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 100
          description: 'Identificador único da competência, em texto simples (ex.:
            "empresa-exemplo"). Não pode repetir.'
          example: exemplo-1
        title:
          type: string
          minLength: 1
          maxLength: 160
          description: Nome da competência.
          example: Automação de testes
        description:
          type: string
          minLength: 1
          maxLength: 2000
          description: Explicação curta da competência.
          example: Criação de testes automatizados para web e API.
        tags:
          type: array
          maxItems: 50
          description: Palavras-chave ligadas à competência.
          items:
            type: string
            minLength: 1
            maxLength: 80
          example:
            - Playwright
            - TypeScript
      title: CompetencyItem
      example:
        id: automacao-testes
        title: Automação de testes
        description: Criação de testes automatizados para web e API.
        tags:
          - Playwright
          - TypeScript
    def-5:
      type: object
      description: Um projeto exibido no portfólio.
      required:
        - id
        - name
        - description
        - stack
        - href
        - status
        - visibility
        - linked
      additionalProperties: false
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 100
          description: 'Identificador único do projeto, em texto simples (ex.:
            "empresa-exemplo"). Não pode repetir.'
          example: exemplo-1
        name:
          type: string
          minLength: 1
          maxLength: 160
          description: Nome do projeto.
          example: Projeto Exemplo
        description:
          type: string
          minLength: 1
          maxLength: 3000
          description: O que o projeto faz.
          example: Aplicação de demonstração com testes automatizados.
        stack:
          type: array
          maxItems: 50
          description: Tecnologias usadas.
          items:
            type: string
            minLength: 1
            maxLength: 80
          example:
            - React
            - TypeScript
            - Fastify
        href:
          type: string
          format: uri
          description: Link do projeto (repositório ou site).
          example: https://github.com/exemplo/projeto-exemplo
        status:
          type: string
          minLength: 1
          maxLength: 80
          description: Situação do projeto.
          example: Em andamento
        visibility:
          type: string
          enum:
            - public
            - private
          description: "`public` aparece para todos; `private` fica restrito."
          example: public
        linked:
          type: boolean
          description: "`true` se o projeto tem link clicável."
          example: true
      title: Project
      example:
        id: projeto-exemplo
        name: Projeto Exemplo
        description: Aplicação de demonstração com testes automatizados.
        stack:
          - React
          - TypeScript
          - Fastify
        href: https://github.com/exemplo/projeto-exemplo
        status: Em andamento
        visibility: public
        linked: true
    def-6:
      type: object
      description: Um item da trajetória profissional.
      required:
        - id
        - period
        - company
        - role
        - summary
        - details
      additionalProperties: false
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 100
          description: 'Identificador único da experiência, em texto simples (ex.:
            "empresa-exemplo"). Não pode repetir.'
          example: exemplo-1
        period:
          type: string
          minLength: 1
          maxLength: 120
          description: Período em texto livre.
          example: 2024 — 2025
        company:
          type: string
          minLength: 1
          maxLength: 160
          description: Nome da empresa.
          example: Empresa Exemplo
        role:
          type: string
          minLength: 1
          maxLength: 160
          description: Cargo exercido.
          example: QA Analyst
        summary:
          type: string
          minLength: 1
          maxLength: 1600
          description: Resumo em uma frase.
          example: Atuação com testes funcionais e automação.
        details:
          type: string
          minLength: 1
          maxLength: 5000
          description: Detalhes do que foi feito.
          example: Execução de testes, análise de falhas e automação com Playwright.
      title: ExperienceItem
      example:
        id: empresa-exemplo
        period: 2024 — 2025
        company: Empresa Exemplo
        role: QA Analyst
        summary: Atuação com testes funcionais e automação.
        details: Execução de testes, análise de falhas e automação com Playwright.
    def-7:
      type: object
      description: Um curso ou formação.
      required:
        - id
        - title
        - institution
        - status
      additionalProperties: false
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 100
          description: 'Identificador único da formação, em texto simples (ex.:
            "empresa-exemplo"). Não pode repetir.'
          example: exemplo-1
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Nome do curso ou formação.
          example: Curso de Teste de Software
        institution:
          type: string
          minLength: 1
          maxLength: 160
          description: Instituição de ensino.
          example: Instituição Exemplo
        status:
          type: string
          minLength: 1
          maxLength: 80
          description: Situação.
          example: Concluído
        period:
          type: string
          maxLength: 120
          description: Período (opcional).
          example: "2023"
        description:
          type: string
          maxLength: 3000
          description: Descrição (opcional).
          example: Fundamentos de testes, técnicas e boas práticas.
        href:
          type: string
          format: uri
          description: Link do curso ou certificado (opcional).
          example: https://exemplo.com/curso
      title: EducationItem
      example:
        id: curso-exemplo
        title: Curso de Teste de Software
        institution: Instituição Exemplo
        status: Concluído
        period: "2023"
        description: Fundamentos de testes, técnicas e boas práticas.
        href: https://exemplo.com/curso
    def-8:
      type: object
      description: Um destaque ou conquista.
      required:
        - id
        - title
        - description
      additionalProperties: false
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 100
          description: 'Identificador único do destaque, em texto simples (ex.:
            "empresa-exemplo"). Não pode repetir.'
          example: exemplo-1
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Título do destaque.
          example: Redução de falhas em produção
        description:
          type: string
          minLength: 1
          maxLength: 3000
          description: Explicação do destaque.
          example: Testes automatizados reduziram erros após cada entrega.
        href:
          type: string
          format: uri
          description: Link relacionado (opcional).
          example: https://exemplo.com/case
        label:
          type: string
          maxLength: 100
          description: Texto do botão do link (opcional).
          example: Ver case
      title: HighlightItem
      example:
        id: destaque-exemplo
        title: Redução de falhas em produção
        description: Testes automatizados reduziram erros após cada entrega.
        href: https://exemplo.com/case
        label: Ver case
    def-9:
      type: object
      description: Aparência geral do portfólio.
      required:
        - theme
        - language
        - accent
        - intensity
        - spacing
        - pageFormat
        - footerText
      additionalProperties: false
      properties:
        theme:
          type: string
          enum:
            - dark
            - light
          description: "Tema: `dark` (escuro) ou `light` (claro)."
          example: dark
        language:
          type: string
          enum:
            - pt
            - en
            - es
          description: "Idioma: `pt`, `en` ou `es`."
          example: pt
        accent:
          type: string
          pattern: ^#[0-9A-Fa-f]{6}$
          description: Cor de destaque em hexadecimal, com 6 dígitos.
          example: "#7C3AED"
        intensity:
          type: number
          minimum: 0
          maximum: 100
          description: Intensidade dos efeitos visuais (0 a 100).
          example: 60
        spacing:
          type: number
          minimum: 0
          maximum: 100
          description: Espaçamento entre elementos (0 a 100).
          example: 50
        pageFormat:
          type: string
          enum:
            - balanced
            - editorial
            - panoramic
          description: "Formato da página: `balanced`, `editorial` ou `panoramic`."
          example: balanced
        footerText:
          type: string
          maxLength: 500
          description: Texto do rodapé (pode ficar vazio).
          example: © 2025 Maria Exemplo
      title: Appearance
      example:
        theme: dark
        language: pt
        accent: "#7C3AED"
        intensity: 60
        spacing: 50
        pageFormat: balanced
        footerText: © 2025 Maria Exemplo
...
    def-13:
      type: object
      description: Um assunto que o assistente sabe responder.
      required:
        - id
        - title
        - category
        - keywords
        - content
        - enabled
      additionalProperties: false
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 100
          description: 'Identificador único do contexto, em texto simples (ex.:
            "empresa-exemplo"). Não pode repetir.'
          example: exemplo-1
        title:
          type: string
          minLength: 1
          maxLength: 160
          description: Título do assunto.
          example: Como entrar em contato
        category:
          type: string
          minLength: 1
          maxLength: 100
          description: Categoria.
          example: Contato
        keywords:
          type: array
          maxItems: 50
          description: Palavras que ativam este assunto.
          items:
            type: string
            minLength: 1
            maxLength: 80
          example:
            - contato
            - email
        content:
          type: string
          minLength: 1
          maxLength: 5000
          description: Texto usado na resposta.
          example: O contato pode ser feito pelo e-mail ou LinkedIn do portfólio.
        enabled:
          type: boolean
          description: "`true` = o assistente usa este assunto."
          example: true
      title: AssistantContext
      example:
        id: contato
        title: Como entrar em contato
        category: Contato
        keywords:
          - contato
          - email
        content: O contato pode ser feito pelo e-mail ou LinkedIn do portfólio.
        enabled: true
    def-14:
      type: object
      description: Configuração do assistente do portfólio.
      required:
        - greeting
        - suggestions
        - contexts
      additionalProperties: false
      properties:
        greeting:
          type: string
          minLength: 1
          maxLength: 1200
          description: Mensagem de boas-vindas.
          example: Olá! Pergunte sobre projetos, experiência ou contato.
        suggestions:
          type: array
          maxItems: 20
          description: Perguntas sugeridas ao visitante.
          items:
            type: string
            minLength: 1
            maxLength: 120
          example:
            - Quais são os projetos?
            - Como entrar em contato?
        contexts:
          type: array
          maxItems: 100
          description: Assuntos que o assistente conhece.
          items:
            $ref: "#/components/schemas/def-13"
      title: AssistantConfig
      example:
        greeting: Olá! Pergunte sobre projetos, experiência ou contato.
        suggestions:
          - Quais são os projetos?
          - Como entrar em contato?
        contexts:
          - id: contato
            title: Como entrar em contato
            category: Contato
            keywords:
              - contato
              - email
            content: O contato pode ser feito pelo e-mail ou LinkedIn do portfólio.
            enabled: true
...
    def-17:
      type: object
      description: Portfólio completo. Cada bloco tem endpoint próprio; veja os
        schemas de cada um.
      required:
        - profile
        - hero
        - about
        - competencies
        - projects
        - experience
        - education
        - highlights
        - contentBlocks
        - appearance
        - assistant
        - editor
      additionalProperties: false
      properties:
        profile:
          $ref: "#/components/schemas/def-1"
        hero:
          $ref: "#/components/schemas/def-2"
        about:
          $ref: "#/components/schemas/def-3"
        competencies:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/def-4"
        projects:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/def-5"
        experience:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/def-6"
        education:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/def-7"
        highlights:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/def-8"
        contentBlocks:
          type: object
          description: Textos editáveis. Chave = nome do texto, valor = conteúdo.
          additionalProperties:
            type: string
            maxLength: 10000
        appearance:
          $ref: "#/components/schemas/def-9"
        assistant:
          $ref: "#/components/schemas/def-14"
        editor:
          $ref: "#/components/schemas/def-16"
      title: PortfolioDocument
    def-18:
      type: object
      description: Portfólio completo + controle de versão.
      required:
        - revision
        - updatedAt
        - data
      properties:
        revision:
          type: integer
          minimum: 1
          description: Versão atual dos dados. Use em `expectedRevision` ao salvar.
          example: 5
        updatedAt:
          type: string
          format: date-time
          description: Data da última alteração.
          example: 2025-01-15T12:00:00.000Z
        data:
          $ref: "#/components/schemas/def-17"
      title: PortfolioSnapshot
    def-19:
      type: object
      description: "Corpo de escrita: versão atual + os dados novos."
      required:
        - expectedRevision
        - data
      additionalProperties: false
      properties:
        expectedRevision:
          type: integer
          minimum: 1
          description: Versão atual dos dados. Use o valor `revision` retornado pelo
            último GET.
          example: 4
        data:
          $ref: "#/components/schemas/def-1"
      title: ProfileWrite
      example:
        expectedRevision: 4
        data:
          name: Maria Exemplo
          area: Qualidade de Software
          role: QA Analyst
          level: Pleno
          company: Empresa Exemplo
          cep: 00000-000
          address: Rua Exemplo, 123
          location: São Paulo, SP
          email: maria@exemplo.com
          whatsapp: +55 11 90000-0000
          github: https://github.com/exemplo
          linkedin: https://www.linkedin.com/in/exemplo
...
paths:
...
  /api/v1/profile:
    get:
      summary: Mostra o perfil
      tags:
        - Profile
      description: |-
        🔓 **Público** — não precisa de login.

        Retorna nome, cargo, localização, contatos, links e foto.
      responses:
        "200":
          description: Perfil atual.
          content:
            application/json:
              schema:
                description: Perfil atual.
                type: object
                required:
                  - revision
                  - updatedAt
                  - data
                properties:
                  revision:
                    type: integer
                    description: Versão atual dos dados. Use em `expectedRevision` ao salvar.
                    example: 4
                  updatedAt:
                    type: string
                    format: date-time
                    description: Data da última alteração.
                    example: 2025-01-15T12:00:00.000Z
                  data:
                    $ref: "#/components/schemas/def-1"
    put:
      summary: Atualiza o perfil
      tags:
        - Profile
      description: >-
        🔒 **Requer Bearer Token** — faça login em `POST /api/v1/auth/login` e
        cole o `accessToken` no botão **Authorize**.


        Troca nome, cargo, localização, contatos e links. Envie todos os campos
        obrigatórios de `data`. Antes de enviar, consulte o GET correspondente e
        use o valor `revision` retornado em `expectedRevision`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/def-19"
      security: *a1
      responses:
        "200":
          description: Salvo. Retorna o portfólio completo com a nova `revision`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-18"
        "400":
          description: "Dados inválidos: algum campo está faltando ou fora do formato."
          content:
            application/json:
              schema:
                description: "Dados inválidos: algum campo está faltando ou fora do formato."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: VALIDATION_ERROR
                  message:
                    type: string
                    description: Explicação do erro.
                    example: A requisição não atende ao contrato da API.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "401":
          description: "Sem login: envie o Bearer Token (botão Authorize)."
          content:
            application/json:
              schema:
                description: "Sem login: envie o Bearer Token (botão Authorize)."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: UNAUTHORIZED
                  message:
                    type: string
                    description: Explicação do erro.
                    example: Autenticação obrigatória ou token inválido.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "409":
          description: "Conflito: `expectedRevision` está desatualizado (ou o `id` já
            existe). Faça um novo GET e tente de novo."
          content:
            application/json:
              schema:
                description: "Conflito: `expectedRevision` está desatualizado (ou o `id` já
                  existe). Faça um novo GET e tente de novo."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: REVISION_CONFLICT
                  message:
                    type: string
                    description: Explicação do erro.
                    example: O portfólio foi alterado depois que esta edição começou. Recarregue a
                      revisão atual antes de salvar.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
                    example:
                      expectedRevision: 3
                      currentRevision: 4
  /api/v1/profile/photo:
    put:
      summary: Atualiza a foto do perfil
      tags:
        - Profile
      description: >-
        🔒 **Requer Bearer Token** — faça login em `POST /api/v1/auth/login` e
        cole o `accessToken` no botão **Authorize**.


        Troca só a foto, mantendo o resto do perfil. Aceita um link de imagem
        (`https://...`) ou uma imagem em base64 (`data:image/png;base64,...`).
        Antes de enviar, consulte o GET correspondente e use o valor `revision`
        retornado em `expectedRevision`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - expectedRevision
                - dataUrl
              additionalProperties: false
              properties:
                expectedRevision:
                  type: integer
                  minimum: 1
                  description: Versão atual dos dados. Use o valor `revision` retornado pelo
                    último GET.
                  example: 4
                dataUrl:
                  type: string
                  minLength: 1
                  maxLength: 3000000
                  description: Link da imagem ou imagem em base64
                    (`data:image/png|jpeg|jpg|webp;base64,...`).
                  anyOf:
                    - format: uri
                    - pattern: ^data:image/(png|jpeg|jpg|webp);base64,
                  example: https://exemplo.com/foto.jpg
            example:
              expectedRevision: 4
              dataUrl: https://exemplo.com/foto.jpg
      security: *a1
      responses:
        "200":
          description: Salvo. Retorna o portfólio completo com a nova `revision`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-18"
        "400":
          description: "Dados inválidos: algum campo está faltando ou fora do formato."
          content:
            application/json:
              schema:
                description: "Dados inválidos: algum campo está faltando ou fora do formato."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: VALIDATION_ERROR
                  message:
                    type: string
                    description: Explicação do erro.
                    example: A requisição não atende ao contrato da API.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "401":
          description: "Sem login: envie o Bearer Token (botão Authorize)."
          content:
            application/json:
              schema:
                description: "Sem login: envie o Bearer Token (botão Authorize)."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: UNAUTHORIZED
                  message:
                    type: string
                    description: Explicação do erro.
                    example: Autenticação obrigatória ou token inválido.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "409":
          description: "Conflito: `expectedRevision` está desatualizado (ou o `id` já
            existe). Faça um novo GET e tente de novo."
          content:
            application/json:
              schema:
                description: "Conflito: `expectedRevision` está desatualizado (ou o `id` já
                  existe). Faça um novo GET e tente de novo."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: REVISION_CONFLICT
                  message:
                    type: string
                    description: Explicação do erro.
                    example: O portfólio foi alterado depois que esta edição começou. Recarregue a
                      revisão atual antes de salvar.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
                    example:
                      expectedRevision: 3
                      currentRevision: 4
...
servers:
  - url: http://localhost:3333
    description: Desenvolvimento local
...